ADD: resetting password
I pity the more sober me who will undoubtebly wonder what the fuck I did
and why I did it. He will spend hours tracing the changes and additions
I made and why it had to be so. He will suffer hours of looking through
documentation that makes as much sense to him as it does to a rock, as
he would not know where to begin or what he is looking for. Good luck,
son.